refactor: divide entities cmd-handlers.ts business rules
